le project wve

+++ admin/orders/delete_order.php
@@ -1,13 +1,20 @@
 <?php
 include '../../config/db_conn.php';
 
-$id = $_GET["id"];
-$sql = "DELETE FROM orders WHERE id=$id";
 
-if (mysqli_query($conn, $sql)) {
-    header("Location: index_order.php?msg=Commande supprimée avec succès");
-    exit();
+if (isset($_GET["id"])) {
+    $id = intval($_GET["id"]);
+
+    
+    $sql = "DELETE FROM orders WHERE id=$id";
+
+    if (mysqli_query($conn, $sql)) {
+        header("Location: index_order.php?msg=Commande supprimée avec succès");
+        exit();
+    } else {
+        echo "Erreur : " . mysqli_error($conn);
+    }
 } else {
-    echo "Erreur : " . mysqli_error($conn);
+    echo "Aucun ID fourni pour la commande.";
 }
 ?>
